+++ console2/src/components/molecules/RepositoryActionDropdown/index.tsx
@@ -25,6 +25,7 @@ import { ConcordKey } from '../../../api/common';
 import {
     DeleteRepositoryPopup,
     RefreshRepositoryPopup,
+    RepositoryTriggersPopup,
     StartRepositoryPopup
 } from '../../organisms';
 
@@ -49,6 +50,18 @@ class RepositoryActionDropdown extends React.PureComponent<ExternalProps> {
                             <Dropdown.Item onClick={onClick}>
                                 <Icon name="play" color="blue" />
                                 <span className="text">Run</span>
+                            </Dropdown.Item>
+                        )}
+                    />
+
+                    <RepositoryTriggersPopup
+                        orgName={orgName}
+                        projectName={projectName}
+                        repoName={repoName}
+                        trigger={(onClick) => (
+                            <Dropdown.Item onClick={onClick}>
+                                <Icon name="lightning" />
+                                <span className="text">Triggers</span>
                             </Dropdown.Item>
                         )}
                     />